requirements.txt
discord.py==2.4.0
openai==1.54.4
python-dotenv==1.0.1